Update GPIO worksheet and add PDFs
Lots of changes here. The GPIO worksheet now includes Spence's traffic
lights wiring and the button wiring has been updated to include the
traffic lights in case people want to play with that.

The repo is also slightly restructured: PDF output is now included the
sub-directories for the convenience of people who just want those. The
template setup instructions are split out into CONTRIBUTING, and the
README has been updated with thumbnails as per #1 so people can easily
browse the content of the repo.